Re-add master table script for pivoted format
This file was originally master_table.py, but that script now generates
the main melted format. This additional script is still dependent on
functions from master_table.py.

pivot_master_table.py was supposed to do do this more directly by
having the melted format as a prerequisite, but I was not able to
figure out how to do that in time, so I’m just putting this script back
in.

#11